Move tests incorrectly tagged as flaky to main suite
