+++ tests/cli_tests/flaky___flaky_test.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"math"
 	"os"
-	"path"
 	"path/filepath"
 	"regexp"
 	"strconv"
@@ -1003,84 +1002,5 @@ func Test___FlakyScenariosUpdateScSettings(t *testing.T) {
 		// test transaction to verify chain is still working
 		output, err = executeFaucetWithTokens(t, configPath, 1)
 		require.Nil(t, err, "faucet execution failed", strings.Join(output, "\n"))
-	})
-
-	t.Run("delete existing file with thumbnail should work", func(t *testing.T) {
-		t.Parallel()
-
-		allocationID := setupAllocation(t, configPath)
-		defer createAllocationTestTeardown(t, allocationID)
-
-		remotepath := "/"
-		filesize := int64(1 * KB)
-
-		thumbnail := "upload_thumbnail_test.png"
-		thumbnailPath := "./tmp/" + thumbnail
-		output, err := cliutils.RunCommandWithoutRetry("wget https://en.wikipedia.org/static/images/project-logos/enwiki-2x.png -O " + thumbnailPath)
-		require.Nil(t, err, "Failed to download thumbnail png file: ", strings.Join(output, "\n"))
-		defer func() {
-			os.Remove(thumbnailPath)
-		}()
-
-		filename := generateRandomTestFileName(t)
-		err = createFileWithSize(filename, filesize)
-		require.Nil(t, err)
-
-		output, err = uploadFile(t, configPath, map[string]interface{}{
-			"allocation":    allocationID,
-			"remotepath":    remotepath,
-			"localpath":     filename,
-			"thumbnailpath": thumbnailPath,
-		}, true)
-		require.Nil(t, err, strings.Join(output, "\n"))
-		require.Len(t, output, 2)
-
-		fname := filepath.Base(filename)
-		remoteFilePath := path.Join(remotepath, fname)
-
-		output, err = deleteFile(t, escapedTestName(t), createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remoteFilePath,
-		}), true)
-		require.Nil(t, err, strings.Join(output, "\n"))
-		require.Len(t, output, 1)
-		require.Equal(t, fmt.Sprintf("%s deleted", remoteFilePath), output[0])
-
-		output, err = listFilesInAllocation(t, configPath, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		}), true)
-		require.Nil(t, err, "List files failed", err, strings.Join(output, "\n"))
-		require.Len(t, output, 1)
-		require.Equal(t, "null", output[0], strings.Join(output, "\n"))
-	})
-
-	t.Run("delete existing root directory should work", func(t *testing.T) {
-		t.Parallel()
-
-		allocationID := setupAllocation(t, configPath)
-		defer createAllocationTestTeardown(t, allocationID)
-
-		remotepath := "/"
-		filesize := int64(1 * KB)
-		generateFileAndUpload(t, allocationID, remotepath, filesize)
-
-		output, err := deleteFile(t, escapedTestName(t), createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, strings.Join(output, "\n"))
-		require.Len(t, output, 1)
-		require.Equal(t, fmt.Sprintf("%s deleted", remotepath), output[0])
-
-		output, err = listFilesInAllocation(t, configPath, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		}), true)
-		require.Nil(t, err, "List files failed", err, strings.Join(output, "\n"))
-		require.Len(t, output, 1)
-		require.Equal(t, "null", output[0], strings.Join(output, "\n"))
 	})
 }
